move vhs network call to separate file
